Adapted a change that apparently caused windows not to compile all the code. Removed a few more windows warnings from the ATerm library. Extended 3rd party README
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8515 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/aterm/byteencoding.h
@@ -13,7 +13,7 @@ size_t BEserializeMultiByteInt(int i, char *c);
 /* void BEserializeDouble(double d, char *c); */
 
 
-size_t BEdeserializeMultiByteInt(char *c, size_t *i);
+int BEdeserializeMultiByteInt(char *c, size_t *i);
 
 /* double BEdeserializeDouble(char *c); */
